Make utils for loading ycbv scenes (#179)
